upsdate

+++ 23-PizzaAPI_JWT/src/controllers/auth.js
@@ -111,6 +111,45 @@ module.exports = {
     }
   },
 
+  refresh: async (req, res) => {
+    /*
+            #swagger.tags = ["Authentication"]
+            #swagger.summary = "JWT: Refresh"
+            #swagger.description = 'Refresh token.'
+        */
+    const refreshToken = req.body?.bearer?.refresh;
+    if (refreshToken) {
+      const refreshData = await jwt.verify(
+        refreshToken,
+        process.env.REFRESH_KEY
+      );
+      // console.log(refreshData)
+      if (refreshData) {
+        const user = await User.findOne({ _id: refreshData.id }); // refreshData icinde gelen id ile karsilastiriyoruz.
+        if (user && user.password == refreshData.password) {
+          res.status(200).send({
+            error: false,
+            bearer: {
+              access: jwt.sign(user.toJSON(), process.env.ACCESS_KEY, {
+                //veri mongoose  object olarak gelir ve JSON object formatina ceviriyoruz.
+                expiresIn: process.env?.ACCESS_EXP || "30m",
+              }),
+            },
+          });
+        } else {
+          res.errorStatusCode = 401;
+          throw new Error("Wrong id or password.");
+        }
+      } else {
+        res.errorStatusCode = 401;
+        throw new Error("JWT refresh data is wrong.");
+      }
+    } else {
+      res.errorStatusCode = 401;
+      throw new Error("Please enter bearer.refresh");
+    }
+  },
+
   logout: async (req, res) => {
     /*
             #swagger.tags = ["Authentication"]
@@ -120,12 +159,20 @@ module.exports = {
 
     const auth = req.headers?.authorization; // Token ...tokenKey...
     const tokenKey = auth ? auth.split(" ") : null; // ['Token', '...tokenKey...']
-    const result = await Token.deleteOne({ token: tokenKey[1] });
 
-    res.send({
-      error: false,
-      message: "Token deleted. Logout was OK.",
-      result,
-    });
+    if (tokenKey[0] == "Token") {
+      const result = await Token.deleteOne({ token: tokenKey[1] });
+      res.send({
+        error: false,
+        message: "Token deleted. Logout was OK.",
+        result,
+      });
+    } else {
+      // jwt de logout a ihtiyac yoktur.
+      res.send({
+        error: false,
+        message: "JWT: No need any process for logout.",
+      });
+    }
   },
 };
